% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact-rra.R
\name{compute_combinations_table}
\alias{compute_combinations_table}
\title{Compute a matrix of exact per-rank RRA statistics using combinatorics.}
\usage{
compute_combinations_table(size, max_rank)
}
\description{
Helper function for computing a table F of shape (M, N), where F[m, n] is the
number of ways in which M ranks can be chosen from {1, 2, ..., N} such that
at least m of them have a rank equal or lower to n.
}
