% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact-rra.R
\name{compute_pbeta_table}
\alias{compute_pbeta_table}
\title{Functions for computing exact RRA P values
Compute a matrix of per-rank RRA statistics by approximating relative ranks
as uniform distribution variables.}
\usage{
compute_pbeta_table(size, max_rank, log_p = F)
}
\arguments{
\item{size}{Number of ranks to be drawn, i.e. number of rows in the output
matrix.}

\item{max_rank}{Total number of available ranks to be drawn from, i.e. the
number of columns in the output matrix.}

\item{log_p}{Whether to return log-probabilities.}
}
\value{
A M * N probability lookup matrix.
}
\description{
Compute a lookup table F of shape (M, N), where M is the total number of
ranks to be drawn from ranks {1, 2, ..., N} without replacement and F[m, n]
indicates probability that out of a set of M randomly selected ranks from, at
least m of the selected ranks have a rank of n or lower. This probability is
computed by approximating the relative ranks n / N using independent uniform
distributions.
}
\details{
In other words,
F[m, n] = pbeta(n / N, 1, M + 1 - m), where
1 <= m <= M and 1 <= m <= N.
}
