% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact-rra.R
\name{.exact_rra_pval_from_score_mat}
\alias{.exact_rra_pval_from_score_mat}
\title{Helper function for computing RRA P value from a score matrix.}
\usage{
.exact_rra_pval_from_score_mat(score_mat, cutoffs)
}
\arguments{
\item{score_mat}{A M * N score matrix.}

\item{cutoffs}{Significance cutoffs to be used for the ranks.}
}
\value{
P values corresponding to each provided cutoff.
}
\description{
Computes the exact probability of randomly choosing M of ordered ranks
(r_1, ..., r_M) that satisfy the following conditions. score_mat is a matrix
of shape (M, N).
\enumerate{
\item r_1 <= r_2 <= ... <= r_M.
\item min(score_mat\link{m, r_m}) > cutoff for all m in {1, ..., M}.
}
}
