% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact-rra.R
\name{.stat_table}
\alias{.stat_table}
\alias{.stat_table_memoized}
\title{Helper functions for generating a memoized order statistic table of shape
(size, max_rank).}
\usage{
.stat_table(size, max_rank, stat = c("beta", "comb"))

.stat_table_memoized(size, max_rank, stat = c("beta", "comb"))
}
\description{
.stat_table_memoized() is a memoised version of .stat_table().
}
\section{Functions}{
\itemize{
\item \code{.stat_table_memoized}: 
}}

