% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact-rra.R
\name{exact_rra_pval}
\alias{exact_rra_pval}
\title{Compute the exact RRA P value for a rank vector (r_1, ..., r_m).}
\usage{
exact_rra_pval(ranks, max_rank, stat_method = c("beta", "comb"), memoize = T)
}
\arguments{
\item{ranks}{A vector of potentially unordered ranks. NAs are not allowed.}

\item{max_rank}{The maximum value a rank can have, i.e. the total number of
ranked values in the experiments.}

\item{stat_method}{If "beta", then the test statistic is obtained by via
k'th order statistics by approximating normalised ranks as the uniform
distribution, which follows beta distribution. If "comb", then the test
statistic is obtained exactly through combinatorics.}

\item{memoize}{Whether to memoize the computed probability lookup tables.}
}
\value{
A vector of the test statistic, the P value and the index in ranks
that produced the final test statistic.
}
\description{
The robust rank aggregation statistic (RRA) is defined as min(stat(i, r_i)).
This function computes the exact P value for an observed ranks vector r =
(r_1, ..., r_m) under the null hypothesis that the ranks are drawn uniformly
from ranks 1 ... n (without replacement).
}
