% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_binary_search}
\alias{matrix_binary_search}
\title{Binary search for each row in two-way monotonous matrix score_mat}
\usage{
matrix_binary_search(score_mat, cutoff)
}
\arguments{
\item{score_mat}{A M * N matrix of values to be searched.}

\item{cutoff}{The cutoff value.}
}
\value{
A vector J, where value J\link{m} is such that score_mat\link{m, n} <= cutoff
for n <= J\link{m} and score_mat\link{m, n} > cutoff for n > J\link{m}. Note that values in
J are 1-based.
}
\description{
Matrix \code{score_mat} is a M * N matrix, which is monotonous in two ways in the
sense that score_mat\link{m, n} < score_mat\link{m, n + 1} and score_mat\link{m, n} >
score_mat\link{m + 1, n} for each 1 <= m <= M and 1 <= n <= N.
}
